Refactor to allow for testing
