+++ .editorconfig
@@ -8,5 +8,8 @@ charset = utf-8
 insert_final_newline = true
 trim_trailing_whitespace = true
 
+[*.go]
+indent_style = tab
+
 [Makefile]
 indent_style = tab
