+++ Makefile
@@ -13,8 +13,8 @@ test:
 build/$(BINARY)-windows-amd64.exe: $(SRC)
 	mkdir -p build
 	GOOS=windows GOARCH=amd64 \
-	go build -o build/$(BINARY)-windows-amd64.exe ./cmd/
+	go build -o build/$(BINARY)-windows-amd64.exe ./cmd/plugin
 
 .PHONY: clean
 clean:
-	-rm -rf build/
+	-rm -rf build/*
