--- cmd/main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/docker/machine/libmachine/drivers/plugin"
-	"github.com/pecigonzalo/docker-machine-vmwareworkstation"
-)
-
-func main() {
-	plugin.RegisterDriver(vmwareworkstation.NewDriver("", ""))
-}
