+++ cmd/plugin/main.go
@@ -1,0 +1,10 @@
+package plugin
+
+import (
+	"github.com/docker/machine/libmachine/drivers/plugin"
+	driver "github.com/pecigonzalo/docker-machine-vmwareworkstation/pkg/driver"
+)
+
+func main() {
+	plugin.RegisterDriver(driver.NewDriver("", ""))
+}
