+++ go.mod
@@ -2,14 +2,14 @@ module github.com/pecigonzalo/docker-machine-vmwareworkstation
 
 go 1.13
 
+replace github.com/docker/docker => github.com/docker/engine v1.4.2-0.20190717161051-705d9623b7c1
+
 require (
 	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
-	github.com/davecgh/go-spew v1.1.0 // indirect
-	github.com/docker/docker v1.4.2-0.20180330173106-635f359f8b61 // indirect
-	github.com/docker/machine v0.14.0
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/sirupsen/logrus v1.0.5 // indirect
-	github.com/stretchr/testify v1.2.1
-	golang.org/x/crypto v0.0.0-20170704135851-51714a8c4ac1
-	golang.org/x/sys v0.0.0-20180329131831-378d26f46672 // indirect
+	github.com/docker/docker v1.4.2-0.20200309214505-aa6a9891b09c // indirect
+	github.com/docker/engine v1.4.2-0.20200309214505-aa6a9891b09c // indirect
+	github.com/docker/machine v0.16.2
+	github.com/sirupsen/logrus v1.5.0 // indirect
+	github.com/stretchr/testify v1.5.1
+	golang.org/x/crypto v0.0.0-20200323165209-0ec3e9974c59
 )
