+++ pkg/driver/vmrun_linux.go
@@ -1,0 +1,13 @@
+package vmwareworkstation
+
+func detectVMRunCmd() string {
+	return ""
+}
+
+func detectVDiskManCmd() string {
+	return ""
+}
+
+func detectDhcpLeasesPath() string {
+	return ""
+}
